remove sbt-sriracha and enable scala-native test
