+++ .travis.yml
@@ -29,8 +29,8 @@ matrix:
     before_install:
     - sudo apt-get update
     - curl https://raw.githubusercontent.com/scala-native/scala-native/master/scripts/travis_setup.sh | bash -x
-    script:
-    - sbt ++${TRAVIS_SCALA_VERSION}! nativeTest/run
+    env:
+    - TEST_PROJECT="scoptNative"
 
 cache:
   directories:
