+++ native/src/test/scala/scopt/SpecUtilPlatform.scala
@@ -1,0 +1,5 @@
+package scopttest
+
+object SpecUtilPlatform {
+  val envLineSeparator = "\n"
+}
